Make imports of rocksdb agnostic of features
rocksdb type DB is a typedef dependent on whether the
"multi-threaded-cf" feature is enabled during compilation.
The alternative types it aliases exhibit differing APIs as well.

When aptos crates are used as dependencies in an external project
that also uses rocksdb and enables the "multi-threaded-cf" feature,
some crates, most notably aptos-schemadb, fail to compile.

This is a footgun on the part of rocksdb, and the way to avoid it
is to use the specific database types for the chosen threading mode.