move termination logic
move termination logic to `vest` instead of `vest_individual`. This is done
to prevent following from happening:
If 5 shareholders are there and only 2 have non-zero balance when `vest` is called
after these 2 get their vesting due, if `vest_individual` sets the termination flag,
in the loop when `vest_individual` is called for 3rd shareholder assertion will fail
and the whole Tx will abort.