refactor: remove commit_blocks_ext (#14101)
Used to exist to not to change tests,
while in fact most tests pass with the real thing
(no force flushing the bufferred state),
with the exception of some proof verification added later
than the bufferred state change.

Not being able to get state proof when state
is still in the buffer is nonetheless an issue
we need to address if we open APIs to the light client
so I'm removing the hack here and defer that to
further proper treatment. I made the flush interval
small for tests though, to let the verification still
happen occasionally in the integration test.