[move-prover] Fix specification functions referring to generic type info
A specification function which uses `type_info::type_of<A>` where `A` is a type parameter implicitly depends on the Boogie global variable `#0_info`. However, that variable is not added to the Boogie function parameters. This PR adds logic to add additional parameters to specification functions capturing this dependency, if needed.

The issue seem to have been present since we introduced TypeInfo verification (so both in compiler v1 and v2) but hasen't surfaced yet as it appears that for framework verification, the problem is not hit.

Tested by a new case in prover `type_reflection.move` test.

Closes #14205