[compiler-v2] Add test cases for a number of reference safety bugs
This adds the reproducing test cases for all the urgent and some high priority bugs related to reference safety. Some of those issues have overlaps, so its good to see all tests in place and the effect of a given change. The tests are marked with the `TODO(#nnn)` code marker for discoverability.

There is also a test for an issue related to v1/v2 used stack size.