[compiler-v2] Fixing bug in stack balancing of file-format generator
Fixes #13912

The `balance_stack_end_of_block` function in function_generator.rs was both dumb and incorrect. It first pushed the argument which is needed for a branch on the stack, then checked if the stack needs flush, and then flushed the value just pushed right away, leading to stupid code like

```

CopyLoc(temp)
StLoc(temp)
CopyLoc(temp)
BrFalse
```

This code is also wrong if `temp` is borrowed since `StLoc(temp)` is not allowed in this case, which was the bytecode verification failure hit by the bug.

The fix simpplifies the stack balance function to ensure `temp` is not flushed unnecessarily.

Existing tests have been migrated from reference-safety to file-format-generator, and a new test has been added, reflecting a regression first caused by the fix.