[compiler-v2] Do not pop non-droppables on stack flush
Fixes #14233

On stack flush we must not pop values which cannot be dropped. Rather they should be stored in temps. This faciltates scenarios where an `abort` happens before the end of a block. Without this, we pop the value before the abort, leading to a bytecode verification error.

There is still an issue with a missing warning, added #14320 for that.