[types] break lock reentrancy in state_key
When removing entries from the registry,
defer drop of the removed entry to after the lock is released.

If the entry refers to another resource on the same registry shard
and that entry gets dropped in turn, the lock in the TwoKeyRegistry
is attempted to be acquired by this thread again, which results in
a deadlock.