Fix potential deadlock in state_key::Entry::drop
If before Entry::drop() calls Registry::maybe_remove, the weak ref under
the same key1 and key2 is replaced, Entry::drop can be called
recursively resulting in a deadlock. The crux is we try to see if the
entry has been replaced by trying to upgrading it to a strong ref, and
it possible that the upgrade is successful and the result becomes a only
ref to the new entry.

This tries to fix it by determining the same thing by seeing if
Weak::strong_count() is 0 instead of upgrading the weak ref.