resolve rust-setup error
I put in the if- condition to speed up iteration on PRs, and it looks
like when the condition is not hit, rust-setup is not happy with it.
Removing to unbreak main.

```
Error:
aptos-labs/aptos-core/main/.github/actions/rust-setup/action.yaml: The
scalar style 'SingleQuoted' on line 23 and column 11 is not valid with
the tag '!startsWith(github.ref,'
Error: System.ArgumentException: Unexpected type '' encountered while
reading 'action manifest root'. The type 'MappingToken' was expected.
   at
   GitHub.DistributedTask.ObjectTemplating.Tokens.TemplateTokenExtensions.AssertMapping(TemplateToken
   value, String objectDescription)
      at
      GitHub.Runner.Worker.ActionManifestManager.Load(IExecutionContext
      executionContext, String manifestFile)
      Error: Failed to load
      aptos-labs/aptos-core/main/.github/actions/rust-setup/action.yaml
```