move maybe_evict_cache() off default rayon pool
This can potentially block the pre_commit_ledger thread while it's
holding the buffered_state lock. Before we reduce the scope of that
lock, I'm moving this part off the defualt pool so that even if the
default pool gets busy (or worse, there are tasks on the threads of the
pool trying to lock the buffered_state) this part doesn't get affected.