Merge TransactionBlockExecutor and VMExecutor into VMBlockExecutor (#15137)
## Description
After latest round of @msmouse's refactors - all implementations of TransactionBlockExecutor are identical (and delegated to DoGetExecutionOutput), so it is unnecessary. On the other hand VMExecutor is actually a block executor (and a hook-up point to native executors) - so renaming it into VMBlockExecutor.

AptosVM was implementing VMExecutor, which seems odd, so separated into new AptosVMBlockExecutor for that.

For native implementations, I needed also single instance of VMBlockExecutor, so added new() method there and &self, with appropriate warning to AptosVM (i.e. production) implementation of it.

@msmouse , @georgemitenkov let me know if this aligns with your planned work. also if there are loader PRs pending, I can land this after those land - so rebase is on me.