[BlockSTM] Lazy (potentially iterable) txns input to BlockSTM (#14568)
* [BlockSTM] Iterable txns input to BlockSTM

Currently BlockSTM takes in a block (vec) of txns and executes them.
This commits adds a capability where we don't need to provide all the
txns in the block upfront, rather provide them as per any desired logic
in the system.

The commit has a default implementation 'DefaultTxnProvider' where all
txns are provided upfront as per current logic, and also a reference
implementation of 'BlockingTxnsProvider' where txns can be provided
after BlockSTM starts execution. This is done by rust's OnceCell<>.

Note: One should be careful while using 'BlockingTxnsProvider' because
if BlockSTM chooses to execute a txn that is not yet provided, then that
thread gets blocked until such a txn is provided. This could lead to
performance degradation.

* Address issues that can arise from spurious wakeups in Condvar

* In BlockingTransaction use OnceCell<> instead of Mutex and Cvar

* Keep BlockingTransaction internal to BlockingTxnProvider

* Remove BlockingTransaction struct; instead use OnceCell directly