[move] Benchmarking historical transactions (#15329)
A tool to benchmark execution of past transactions. The user has to provide first and last
versions of the interval that need to be benchmarked. The tool partitions transactions in
the specified closed interval into blocks, and runs all blocks end-to-end, measuring the
time. During this run, executor is shared (and so are environment and module caches).

There is no commit here, only execution time. For each block, we maintain the state
(read-set estimated from 1 run before the benchmarks) on top of which it should run.
And so, the benchmark just runs in sequence blocks on top of their initial states
(outputs are only used for comparison against the on-chain data).

The tool allows one to override configs to experiment with new features, or with how
the execution would look like without some features. For now, we support:
- enabling features
- disabling features
In the future, we can add more overrides: gas schedule, modules, etc.