[compiler-v2] Make `-v debug` output more concise (#15642)
Output using the `debug!` logging facility should not stay permanent in the code, or at leastr be protected by an additional guard, unless it is useful to work with end users. The default output of `debug!` should overall be useful for users to help understand their problems and avoid complex internal data structures.

This updates some code in the compiler to follow this guideline, as well as refines the docs for logging to reflect this.

This came apparent when trying to use `aptos move prove -v debug` -- the output printed lots of stuff which is probably only relevant for internal debugging.