[move-vm][closures] Types and opcodes for closures (#15667)
* [move-vm][closures] Types and opcodes for closures

This PR implements the extensions needed in the file format for representing closures:

- The type (SignatureToken) has a new variant `Function(args, result, abilities)` to represent a function type
- The opcodes are extendeed by operations `ClosPack`, `ClosPackGeneric`, and `ClosEval`

This supports bit masks for specifyinng which arguments of a function are captured when creating a closure.

Bytecode verification is extended to support the new types and opcodes. The implementation of consistency, type, and reference safety should be complete. What is missing are:

- File format serialization
- Interpreter and value representation
- Value serialization
- Connection of the new types with the various other type representations

* Addressing reviewer comments

* Rename UNIMPLEMENTED_FEATURE into UNIMPLEMENTED_FUNCTIONALITY to avoid confusion with feature flags.

* Add `is_assignable_from` and apply at the right places in type_safety checker