[compiler][closures] Wire compilation down to binary format and fix bugs
This implements closures in the RTL bytecode ("stackless bytecode") and wires them down in the compilation chain to Move bytecode. No actual execution of the codes happens, but bytecode verification.

A few issues found in closure compiler and the bytecode verifier are fixed, as well as missing pieces in the stack extended (like the disassembler).

Existing tests (after modification of sources) now pass also bytecode verification.