[move 2.2][closures] Support function type wrappers in the language
It's awkward to deal with function types for resource indexing directly because of the requirement of annotating abilities. This PR adds two simple features which enables named function types by newtype wrappers.

One can write `struct Operation(|u64|u64) has store` to create a named function type which is technical just a positional struct. The inner function fields inherit the abilities of the struct. Function application is enabled so one can directly write `let oper = Operation(|x|x + 1_); oper(x)`. However, the constructor always need to be called to construct a value.

This is probably the style we will also suggest in the book when dealing with function values as data. It adds clarity to the design, and avoids problems with type inference.