[move][closures] Persistent functions and module lock
One last open problem with function values is identification of whether functions are storable which are imported from another module. Currently, the CompiledModule has no information about whether an imported function is friend or public in the FunctionHandle.

This PR solves this problem, and via some generalization, also two other outstanding problems:

1. A mechanism to store attributes with FunctionHandle and FunctionDefinition is introduced. This is a new field in this structs in file_format which contains a list of attributes.
2. An attribute `Persistent` is introduced. The compiler attaches this attribute to handles derived for public functions, and the logic in the VM checks this attribute to figure whether a function is storable.
3. A user can also attach the attribute via `#[persistent] fun`. This closes a 2nd problem in the function value design, namely that a contract is forced to make functions public to store them, which is a security concern. With the attribute, non-public can be made storable as well. The compatibility checker has been extended to treat a function with this attribute like it would be public.
4. Given that we know have attributes at functions in the VM, a 2nd attibute `#[module_lock]` was introduced. This allows to enable module locking for reentrancy, as it is currently used in native dispatch, on user level. This addresses a concern raised by Ottersec in pre-auditing discussions.

There are multiple tests added for those new features, both transactional as e2e. A ton of base line files are changed because many tests print the list of attributes known to the compiler, which has changed.