[move-vm][closures] Prepare serialization format and resolver for late binding
This sets up closure serialization and resolution for late binding. While the feature is not yet connected to Move code, which happens in the future, setting up this functionality now prevents breaking changes in the serialization format.

Late binding, as used e.g. in FA dispatch, allows to programmatically create function values via their name.  Late binding is almost already supported. The only thing missing is (a) a way to specify the expected type when the function resolves and a runtime check that it does (b) a check that the resolved function is public.

Technically, there is a new `kind` field in serialized function data. That field will also allow us to evolve this data in the future via versioning.

This *also* removes the previous verification of the type layout stored with a closure against a loaded function's type. This check is not really needed since we generally already assume storage consistency. (For instance, we also do not verify the serialized data in a struct matches the types of the struct definition.) Verification is needed for late bound functions. Those two different use cases where previously mixed up, and are now cleanly seperated.