[move][rac] Revising resource access control (#16081)
* [move][rac] Revising resource access control

This PR consolidates RAC with the Move 2 design and solves some open language integration problems. It also makes a few changes to the file format representation steming from that.

The open question was the combination of `acquires` with read/write access clauses in a way compatible with existing code. This PR separates the traditional `acquires` from the read/write clauses.

- From language version 2.2 on, `acquires` is optional. However, if one is provided, it must be precise in the traditional way. For the VM, acquired resources will be inferred (in fact they were already before).
- Independent of acquires, a user can provide `reads` and `writes` annotations. The `acquires` as an access specifier is removed, also in the VM. Write access subsumes now read access.

This also cleans up some test code in the compiler, which leads to a larger diff. Specifically, `LanguageVersion::latest()` is now everywhere the default. However, for checking, we have language version specific directories, and according tests are move into those.

There are more changes to do in the VM in subsequent PRs, but transactional tests pass, so the feature is now working.

Closes #12623

* Addressing reviewer comments, plus significantly simplifying RAC logic by dropping the ambition (for now) to normalize access specifiers.

* Addressing reviewer comments

* Addressing reviewer comments