[move-example] sample of a dynamic dispatch engine in Move (#14132)
This demonstrate a proof of concept dynamci dispatch engine in Move that
might be useful for applications like bridges or swaps. For bridges,
this allows for a bridge to call into user code without having to do the
complicated dance of the application layer knowing the on-chain function
to call and ensuring end-to-end correctness and validation. For aggregated
swaps, this decouples the need to load all dependencies for all swaps at
startup time.

Co-authored-by: Greg Nazario <greg@gnazar.io>