[move][closures] Fixing bug in handling of references in function types (#16152)
When a function type is used in a field, the parameter and result types can be references, but the verifier did not allow this before and treated them as all other inner types. There was also a related bug in the unused param checker of the ccmpiler which did not expect references.

Closes #16147