[compiler] Fixing bug in variance of function types
The variaane parameter was ignored on unification of function types, allowing variance where it shouldn't. Closes #16401.

This also fix a bug in display of function, which had the wrong attempt compacting abilities when defaults are used. This does not work becxause we also nneed to represent empty abilities. Leads to some baseline changes.