[compiler] Ensure that all arguments of lambda are curried (#16474)
* [compiler] Fixing a bug in inferred type defaults

Inferred type defaults where not checked for abilities. Closes #16405

* [compiler] Fixing cyclic type replacement leading to stack overflow

When creating default types from constraints during type finalization we can run into cycles, leading to stack overflow. This adds a visiting set to the replacement algorithm to bail out in this case.

Closes #16435

* Update comment.

* [compiler] Ensure that all arguments of lambda are curried

In order currying to succeed, we must actually use all arguments of the lambda in the curried function.

Closes #16468