block prologue transaction sees latest timestamp
The WriteOpConverter was seeing the timestmap of the previous block.
It used not to be a problem because the timestmap is only used when a
new slot is created, which doesn't happen for the block metadata
transaction.
Now that we intend to use (and later write down) the access timestmap
in addition to the creation timestmap, it should be fixed.