Connect HotState with ReadWriteSummary (#16485)
* move state_slot.rs

* implement get_state_slot() version() for CachedStateView

Give the ability to return StateSlot and the version for CachedStateView, which can be seen in the block stm work loop.

The hot state promotions and evictions are not yet deterministic, serialized or charged, so it's okay for other state views to return fake slots and versions.

* BlockEndInfoExt that carries keys to promote to make hot

Not serialized and stored on-chain for now, not until the hot state is fully deterministic.

Always Empty for now.

* implement BlockHotStateOpAccumulator

* WriteSet carries either WriteOps or HotStateOps

removed iterator implementations on WriteSet so on the existing
callsites it's explicit that we are iterating WriteOps only, not
including HotStateOps

* Storage follows HotStateOp

CachedStateView no longer responsible for calculating hot state
refreshes, as a result (temporarily), fullnodes in apply mode won't
cache read-only keys in the hot state.

* update test_speculative_state_workflow to use HotStateOpAccumulator

* update golden files

WriteSet's debug output changed.

* remove StateSlot::HotVacant::deletion_version

I feel it's not worth it to make it so confusing in order to maintain
the invariance that all the created nodes must appear in the cache.

IMO It's not worth it to maintain the invariance by detecting if nodes are
"actually" changed during a update either.

* fix

* fix debugger

* add counters

* comments

* remove proptest-regressions