[storage] Split hot state into 16 shards
The writes (`StateUpdateRef`s) have already been split into shards. We also
intend to store hot state as 16 shards in persistent storage. So it's more
convenient (and eventually more performant) to make the in-memory representation
16 shards as well.

Right now the max size of each shard is the old max size / 16, but we might
tweak this in the future.