[hot state] Populate hot state promotions
This was originally done in #16485. However, subsequent changes (mostly priority
fee implementation) undid some of the changes, so while
`BlockHotStateOpAccumulator` still computes `slots_to_make_hot`, the result is
never populated to the hot state storage. Therefore, the read-only keys are
never cached.

This change poluates the infomation again by putting it inside `BlockOutput`.

The tricky part is that the read write summary can sometimes be inaccurate. For
example, occasionally a key is written to, but it's not in the write summary,
then the code would generate both a `Modification` op and a `MakeHot` op for
the same key. To be defensive, if `Creation`/`Modification`/`Deletion` is
followed by `MakeHot` on the same key when computing `BatchedStateUpdateRefs`,
we have to discard the `MakeHot`.