[move][transactional tests] Enable tera templates for transactional tests
[Tera](https://keats.github.io/tera/) is a templating language for Rust with similar concepts like Jinja2 (Javascript) or Django (Python). Those include mainly the abilities to define macros and to do template expansion using iterations and conditionals. We have been using Tera extensively as part of the Move prover to generate Boogie output from sttackless bytecode.

This adds template support to transactional test framework. Whenever a file input to TTF starts with (after whitespace) `{{`, the tera templating engine will be called to expand the input. The first line for such a test will typically import macros.

Example:

```
{{ import "macros" as m }}
{% for i in [1,2,3] %}
{% for ty in integer_types %}
//# run
script
fun main()
   ld_{ty} {{i}}
   {{m::expect(ty=ty, val=i)}}
   ret
{% endfor %}
{% endfor %}
```

This feature is still experimental so no documentation yet, and also only used for demonstration in a few examples. I'm not sure whether we want to fully commit on something like this. While it increases flexibility to write transactional tests, it may also make maintenance harder.

Another change in this PR is that the test framework now prints information about each executed command independent of whether failure or not if the new config `echo` is set. This makes it easier to review specifically large amounts of tests resulting from template expansion. Probably want to make this default for all TTF outputs, but not in this PR.