[Hot State] Record read-only promotions in BlockEndInfo
The previous approach in #17136 was to put `to_make_hot: Map<Key, Slot>` in `BlockOutput`, and let
`execute_block` compute and return this information.

It works for normal block execution. But if we are doing state sync and want to re-execute
transactions, it will not work, because it's hard to compute this information given a transaction
chunk. Therefore we still need to record this information in block epilogue transactions and simply
reuse them during re-execution.

For now it's in `TBlockEndInfoExt`, and it's not serialized or persisted until later.