[move] Cross-compiling transactional tests for roundtrip testing decompiler/disassembler
This PR adds a 'cross-compilation' mode of a transactional test. A cross-compiled transactional tests replicates the original test step-by-step, but replacing any source content for `//# publish` or `//# run` commands with a decompiled Move or diassembled MASM. It can then run the cross-compiled test as part of the same transactional test run.

Cross-compilation mode is activated by a TestRunConfig which can mark multiple cross targets, e.g.:

```
 run_test_with_config(
        TestRunConfig::default()
            .cross_compile_into(SyntaxChoice::ASM, true)
            .cross_compile_into(SyntaxChoice::Source, true),
        path,
    )
```

The cross-compiled sources will are placed in a sub-folder relative to the current test named 'round-trip', with ending '.disassembled' or '.decompiled' depending on the target. When execution of the cross-compiled test is choosen, a '.exp' will be placed side-by-side with the cross-compiled source. Notice that exp files cannot be directly compared with the original exp since line numbers between original source and cross-compiled source changed, so manual inspection of those .exp files is recommended.

The cross-compilation features is intended to eventually also allow to port remaining MVIR test over the MASM, and retire the MVIR compiler, to happen in future PRs.

For now the feature has been applied for round-trip testing in `third_party/move/testing-infra/transactional-test-runner` and `third_party/move/tools/move-asm`. Bugs in the diassembler have been fixed as a result of this, all tests for MASM are now successfully roundtripped MASM -> Bytecode -> MASM -> Bytecode. For the decompiler, there seem to be some more issues and the cross compilation feature hasn't been widely used for it yet.