[hot state] Compute promotions only at the end of each block
Currently, the code takes each transaction's reads and writes, computes the
read-only keys to promote, then moves on to the next transaction.

There are a few problems:

- Let's say a transaction has read-only keys A and B. It could happen that A
  gets promoted, but B couldn't because it hits the per block limit. However,
  it's possible that a later transaction actually writes to A, so we could have
  promoted B.
- More importantly, the result may not be deterministic if the order we call
  `add_transaction` is not deterministic (which could happen in parallel
  execution).

Therefore, we simply record the reads and writes in `add_transaction` and only
compute the promotions at the end, once all transactions have been added.
Eviction will need to be computed this way as well.