[Storage] Set cache_index_and_filter_blocks by default
I'm not sure the current configuration makes a lot of sense in terms of memory
management.

`max_open_files = 5000` is quite high. Given that this limit is per database and
we have at least 40~50 of them (16 state kv shards, 16 state merkle shards,
etc), we probably rarely hit this limit. So the files remain open once they are
touched.

In addition, we have `cache_index_and_filter_blocks = false`.

The combination of these means that index/filter blocks also stay in memory
forever once they are loaded, even if they are just used once and never touched
again. Therefore, we can waste quite some memory on unused things. We can verify
this by looking at some full nodes that are running at 30~50GB RSS: the stats
show that `rocksdb.estimate-table-readers-mem` is a major contributor, often
more than 70%. See https://gist.github.com/wqfish/4b2b3bccba6466d31ab7732524967ed6
for an example.

This change puts index/filter blocks in the block cache, so the unused ones can
be evicted when the cache is full. I think for now the numbers are relatively
conservative, and we might able to reduce them further. Also we are using
separate caches for different things (bigger ones for each state kv shard and
smaller ones for everything else), but we can also consider experimenting with
using a single big cache for everything, so it's even easier to manage.

`rocksdb-0.24.0` upgrade is needed, because we need `RocksDB >= 8.7.0` to use
"automatic" variant of `HyperClockCache` that does not require a prior estimate
of the average size of cache entries