[vm] Suppress paranoid checks for trusted code
This introduces the concept of 'trusted code' in the VM. For now this is defined as code at a 'special' address (0x1-0xf), but this can later be extended.

When trusted code is executed, paranoid mode is deactivated. Specifically, for trusted code it can be assumed that the deployed bytecode is not maliciously engineerd but generated by the Move compiler and has been gone through rigerous testing, auditing, and verification steps.