[Storage] Use thread local counters
This change switches majority of the storage counters to the thread local
version.

Not all of them are changed. For a few of them, some code is using the `get` API
or other things, which is a bit trickier for a thread local counter. So I'm not
touching these yet.

Tested by running

```
cargo run --release --package aptos-executor-benchmark \
  -- \
  --execution-threads 32 \
  --block-executor-type native-vm-with-block-stm \
  --transactions-per-sender 1 \
  --block-size 8086 \
  --enable-storage-sharding \
  run-executor \
  --enable-feature NEW_ACCOUNTS_DEFAULT_TO_FA_APT_STORE \
  --enable-feature OPERATIONS_DEFAULT_TO_FA_APT_STORE \
  --transaction-type apt-fa-transfer \
  --transaction-weights 1 \
  --module-working-set-size 1 \
  --main-signer-accounts 20000 \
  --additional-dst-pool-accounts 600000 \
  --data-dir ~/db_with_100M_account \
  --checkpoint-dir ~/checkpoint \
  --blocks 2000
```

before and after this change.

Overall I think we will get maybe ~5% throughput improvement, after maybe a few
more changes to the other components.