[Metrics] Create a thread local version to be used in hot paths
We have lots of counters in the codebase. Often, the way we use them is that we
call `with_label_values` and then record values.

`with_label_values` would need to acquire a read lock, which is usually not
expensive but can be nontrivial in hot paths if there are many threads.

This change creates an alternative which is a thread local version. The caller
would just record the value in the thread local counter/histogram, and flush
periodically. This should reduce the contention and speed things up a little.