[Hot State] Tweak LRU interface and tests
I'd like to redesign a few things in the hot state implementation. More in later
commits. For this one, we change the interface such that `HotStateLRU` would
compute eviction itself based on its own state, rather than taking instructions
from the caller. This is more natural and closer to a normal LRU implementation.