[Hot State] Output read promotions per transaction
This is a somewhat big change from the original design.

Previously, we accumulate all the keys read by transactions in the block, filter
out everything that's also in the write sets, then record the remaining keys
(the read-only ones) as promotions in the block epilogue. We have to record them
in the epilogue, otherwise state-sync by re-executing the transactions won't be
able to reproduce the same result (applying output is fine), unless we execute
exactly one block at a time.

This is a bit awkward for the implementation. On one hand, we need to prepare
the promotions before the epilogue, in order to construct the epilogue
transaction. On the other hand, the epilogue transaction itself also produces
its own write set, which can potentially invalidate (filter out) some read
promotions in the previous transactions. So this creates a bit of circular
dependency. We either modify the epilogue transaction again after it's executed,
or we just ignore the epilogue's output, which means we miss some read
promotions and complicates the state computation. None of the workarounds seem
particularly nice.

This change makes each transaction emit its own read promotions, instead of
aggregating them together per block. It simplifies a few things, mainly:
- Regular block execution and state-sync chunk execution no longer have to deal
  with this differently.
- The circular dependency is gone and we don't even have to treat block epilogue
  differently.

The main downside is that size of the write sets may increase a bit:
- Imagine that all transactions read the same key, then the key will appear in
  all the write sets, instead of once per block.
- If a transaction has a read-only key `A` and the next transaction in the same block writes
  to key `A`, we don't record `A` as a promotion in the original implementation,
  but now we have it.

We can measure the size increase, and if it's significant enough to be a
concern, we can simply do some dedup in the executor once the VM finishes
execution. Basically we will still have the same set of promoted keys, just they
are spread across many transactions instead of all living in the epilogue.