[Hot State] Output read promotions per transaction
This is a somewhat big change from the original design.

Previously, we accumulated all keys read by transactions in a block, filtered
out those that also appeared in write sets, and then recorded the remaining keys
(the read-only ones) as promotions in the block epilogue. We had to record them
in the epilogue; otherwise, state sync via re-execution wouldn’t reproduce the
same result (applying outputs is fine) unless we executed exactly one block at a
time.

This makes the implementation a bit awkward. On the one hand, we need to prepare
the promotions before the epilogue to construct the epilogue transaction. On the
other hand, the epilogue transaction itself produces its own write set, which
can invalidate (filter out) some read promotions from earlier transactions. This
creates a circular dependency: we have to modify the epilogue transaction after
it runs. Or we ignore the epilogue’s output, which causes us to miss some read
promotions and complicate state computation as we can no longer guarantee that
reads and writes do not intersect. Neither workaround is particularly nice.

This change makes each transaction emit its own read promotions instead of
aggregating them together per block. It simplifies several things, mainly:
- Regular block execution and state-sync chunk execution no longer need
  different handling.
- The circular dependency is gone and we don't even have to treat block epilogue
  differently.

The main downside is that size of the write sets will increase:
- Imagine that all transactions read the same key, then the key will appear in
  all the write sets, instead of once per block.
- If a transaction has a read-only key `A` and the next transaction in the same
  block writes to key `A`, we wouldn't record `A` as a promotion in the original
  implementation, but now we have it.

We can measure the size increase, and if it's significant enough to be a
concern, we can simply do some dedup in the executor once the VM finishes
execution. Basically, after this change, we can still have the same set of
promoted keys, just they are now spread across many transactions instead of all
living in the epilogue.

Also, this change implements a small optimization where we do not store
`prev_slot` for the promoted keys. The write set now simply says "these are the
keys to be promoted". This will also reduce the size a bit.