[Hot State] Prepare for having all checkpoint indices in StateUpdateRefs
Right now `StateUpdateRefs` has the most recent checkpoint. In order to compute
hot state, we need to perform eviction at the end of each block. When applying
to state sync, this means that we need to know all the checkpoints and not just
the most recent one.

Also refactored the code a bit such that `extract_retries_and_discards` is
simply responsible for extracting retries and discards, not for computing
`has_reconfig`.