[vm] Supress paranoid checks for trusted code (#17461)
* [vm] Suppress paranoid checks for trusted code

This introduces the concept of 'trusted code' in the VM. For now this is defined as code at a 'special' address (0x1-0xf), but this can later be extended.

This also adds nmew transactional tests.  Adding transactional tests.

We also need to move existing tests out of the default trusted 'domain' (0x1-0xf), so refactoring some test files.

* Addressing reviewer comments.

* Ensuring that during tests, no code is trusted. Also adds a mechanism to have different feature flag sets during testing and in prod.

* Addressing reviewer comments

* Reverting change of feature flags for test genesis.

* Updating calibration values for aptos-move/e2e-benchmark and testsuite/single_node_performance.