[Hot State] Compute LRU in speculative state
Until now we had only been computing hot state (its insertions and evictions)
until a block is committed. However, we will eventually need consensus to agree
on the content of the hot state. This means that we need to compute it when a
block is proposed.

This commit finally implements the computation of LRU insertion and eviction in
the speculative state.

Main changes:
- We now use per-version updates instead of batched updates to run
  `State::update`. This is needed for LRU computation. Batched updates are still
  passed to the function for state usage computation, and we'll see if this can
  be cleaned up a bit later.
- The code in `storage/aptosdb/src/state_store/hot_state.rs` now simply accepts
  the updates from upstream, without computing the LRU on its own.

Added tests and validations.