[vm] Non-generic interpreter function caches (#17591)
For non-generic functions, we can keep FrameTypeCache live per
the whole interpeter session - no need to re-instantiate types
and re-calculate various properties.

Note: non-generic caches can be cross-block, and cocnurrent. It
is not clear doing so brings benefits, as we then need synchronisation
on cache miss. Ideally, we pre-compute such things at load time so
that synchronisation is not needed.