Add sigma proofs to aptos-dkg (#17735)
Adds the sigma protocol trait, with accompanying prove and verify code, and Fiat-Shamir code, and a procedural macro SigmaProtocolWitness which allows for easy trait derivation on the Domain.

fixedbasemsms.rs sets up a trait required to efficiently verify sigma proofs for homomorphisms which are essentially a collection of MSMs. Also has formalism to take the products of such maps, which will be used in the upcoming field PVSS and is demonstrated in the new tests for the Chaum-Pedersen protocol.

In particular, this formalism is now used for the KZG commitment scheme (now housed in a separate pcs folder) in the original DeKART range proof.

fiat-shamir.rs has the following important changes:
- Generalised ScalarProtocol to use both arkworks and blstrs.
- Expanded ScalarProtocol to handle generation from 128-bit chunks.
- Added more functions to the RangeProof trait; most of the new functions are marked with #[allow(dead_code)] and will be used in the next version of DeKART.