[annotator] Avoid constructing and scanning values for tables (#18088)
This adds a new function to the annotator whether a given type is or transitively contains tables. Based on that, the code in `db_v2.rs` avoids creating an annotated value if its known to contain no tables.

Consider `vector<T>`, before this change, indexer would visit any value in the vector scanning for table handles even if its known that `T` doesn't has tables.

The information whether tables are in a type is computed during FatType construction and shared between cached structs, as well as cached in the annotator itself.

One step further then this PR goes, would be move table handle scanning into the annotator and avoid to construct annotated values in the first place, but this here might already have significant effects.