[replay-verify] Use performance images
I noticed that the binary we use to run replay-verify (`aptos-debugger`) is
currently built with `cli` profile. This means that the build optimizes more for
binary size instead of performance.

This change switches the default profile to `performance` and we get quite a bit
of speedup (2h -> 1h27m):
https://github.com/aptos-labs/aptos-core/actions/runs/19448179748/job/55647615373

Not sure if we want to leave an option to allow users to specify a
non-performance image, but the github workflows are kind of annoying to modify
lol.