Build C/C++ code with -march=x86-64-v3
Previously in #17643 we started building Rust code with `-C target-cpu=x86-64-v3`
and got at least a few percent performance improvements. This change will start
applying the same to C/C++ code, e.g. jemalloc, rocksdb, etc. It should improve
the quality of codegen, even if this is a smaller portion of the codebase.