[Hot State] Add separate StateMerkleDb for hot state
In order to compute the hot state root hash for each block (including the
speculative ones), we have to be able to return Merkle proof for any given key
and version. The most straightforward way to do this is to add another instance
of `StateMerkleDb` in `StateStore`.

For now, we always delete and reset the DB on restart. This will give us more
flexibility to make changes.