[Layered Map] Expose inner layers
Previously, in order to find all changes between two `State`s, we could just "diff" the two map layers and we would get all the KVs that are changed.

This is insufficient with tiered storage. For instance, consider the following sequence of events:
1. At the end of block `N`, key `K` is read and promoted to hot and its value is `V0` in both hot state and cold state.
2. At the end of block `N+1`, its value is updated to `V1` in hot state.
3. At the end of block `N+2`, it's evicted and the value in cold state is updated to `V1`.
4. At the end of block `N+3`, it's read and promoted into hot state again.

If we simply look at the `StateDelta` in the current implementation, we'd see the value change in hot state, but miss the change from `V0` to `V1` in the cold state.

So we need to expose all the inner layers in order to accurately compute what changed between two points in time.